requirements.txt
groq>=0.4.0
pydantic>=2.0.0
rich>=13.0.0
click>=8.0.0
python-dotenv>=1.0.0
tiktoken>=0.5.0
aiofiles>=23.0.0
gitpython>=3.1.0